requirements.txt
dominate==2.5.2
matplotlib==3.1.3
torchvision==0.2.1
torch==1.0.0
requests==2.25.1
visdom==0.1.8.9
natsort==7.1.0
scipy==1.5.3
mayavi==4.7.2
numpy==1.19.1
scikit_image==0.17.2
beautifulsoup4==4.10.0
completion==0.1.2
Pillow==9.0.1
skimage==0.0
